feat(monitoring): add structured JSON logging and stage monitoring stack
- Add logstash-logback-encoder for JSON console and file appenders
- Create logback-spring.xml with MDC traceId, symbol, module fields
- Add TraceIdFilter for automatic request correlation
- Configure stage-only monitoring stack (Grafana + Prometheus)
- Add JVM/Spring Boot Grafana dashboard (9 panels)
- Fix pre-existing compile errors (ConcurrentHashMap imports, BigDecimal)
- Update dev-stack.sh with stage-monitoring and logs-json commands

+++ backend/api/src/main/java/com/swingtrade/api/service/SignalEngine.java
@@ -181,7 +181,7 @@ public class SignalEngine {
         signalRepository.save(signalEntity);
 
         logger.info("Generated {} signal for {} on {} (confidence: {}%, warning: {})",
-                    signal.type(), symbol, latestDate, String.format("%.0f", signal.confidence() * 100), warningFlag);
+                    signal.type(), symbol, latestDate, String.format("%.0f", signal.confidence().doubleValue() * 100), warningFlag);
     }
 
     /**
